requirements.txt
aequitas==0.42.0
alabaster==0.7.10
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
anyio==3.6.1
argcomplete==1.9.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
CensusData==1.15.post1
certifi==2017.7.27.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
configparser==3.5.0
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Dickens==1.0.1
distlib==0.3.5
docutils==0.13.1
dominate==2.7.0
entrypoints==0.4
enum34==1.1.6
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
Fiona==1.8.21
flake8==3.3.0
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
fonttools==4.33.3
gcc7==0.0.7
gensim==4.2.0
geopandas==0.10.2
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
imagesize==0.7.1
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.2
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.1
kaleido==0.2.1
kiwisolver==1.4.2
lab==7.1
markdown2==2.3.5
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
millify==0.1.1
mistune==0.8.4
munch==2.5.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
ohio==0.5.0
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
Pebble==4.6.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.7.24
platformdirs==2.5.2
plotly==5.8.2
plumbum==1.6.4
portpicker==1.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==2.8.1
pyproj==3.3.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
reportlab==3.6.11
requests==2.27.1
retrying==1.3.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.4.0
Send2Trash==1.8.0
Shapely==1.8.2
signalled-timeout==1.0.0
simplejson==3.17.6
six==1.10.0
sklearn==0.0
smart-open==6.0.0
sniffio==1.2.0
snowballstemmer==1.2.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.2
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenize-rt==4.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.1
traitlets==5.2.2.post1
trio==0.21.0
trio-websocket==0.9.2
txt2tags==3.7
typing==3.6.1
urllib3==1.26.9
vega==3.6.0
verboselogs==1.7
virtualenv==20.16.2
virtualenv-clone==0.5.7
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.1
widgetsnbextension==3.6.1
wrapt==1.14.0
wsproto==1.1.0
xgboost==1.6.1
xhtml2pdf==0.2.2
